+++ fjs/effects/node/memory/module.ts
@@ -7,7 +7,8 @@
 import { randomUUID } from 'node:crypto'
 import { asyncRun } from '../../module.ts'
 import type { Effect, ToAsyncOperationMap } from '../../types.ts'
-import { asBase, asNominal, type Key, type MemOp } from '../../memory/module.f.ts'
+import { asBase, asNominal } from '../../memory/module.f.mjs'
+import type { Key, MemOp } from '../../memory/types.ts'
 
 export type MemoryOperationMap = ToAsyncOperationMap<MemOp>